Remove abstract class and bug fixes
