+++ PSReadLine/Completion.cs
@@ -136,16 +136,6 @@ namespace Microsoft.PowerShell
             _singleton.CompleteImpl(true);
         }
 
-
-        /// <summary>
-        /// Some summary
-        /// </summary>
-        public static void DynamicHelpBlock(ConsoleKeyInfo? key = null, object arg = null)
-        {
-            bool isF1 = key?.Key == ConsoleKey.F1;
-            _singleton.DynamicHelpBlockImpl(isF1);
-        }
-
         private bool IsConsistentQuoting(Collection<CompletionResult> matches)
         {
             int quotedCompletions = matches.Count(match => IsQuoted(match.CompletionText));
@@ -209,200 +199,6 @@ namespace Microsoft.PowerShell
                 }
             }
             return replacementText;
-        }
-
-        private void DynamicHelpBlockImpl(bool isFullHelp)
-        {
-            if (InViInsertMode())   // must close out the current edit group before engaging menu completion
-            {
-                ViCommandMode();
-                ViInsertWithAppend();
-            }
-
-            int cursor = _singleton._current;
-            string commandName = null;
-            string parameterName = null;
-
-            foreach(var token in _singleton._tokens)
-            {
-                if (token.TokenFlags == TokenFlags.CommandName)
-                {
-                    commandName = token.Extent.Text;
-                }
-
-                var extent = token.Extent;
-                if (extent.StartOffset <= cursor && extent.EndOffset >= cursor)
-                {
-                    if (token.Kind == TokenKind.Parameter)
-                    {
-                        parameterName = ((ParameterToken)token).ParameterName;
-
-                    }
-                    // Possible corner case here when cursor is at the end
-                }
-            }
-
-            if (!String.IsNullOrEmpty(commandName) && isFullHelp)
-            {
-                System.Management.Automation.PowerShell ps = System.Management.Automation.PowerShell.Create(RunspaceMode.CurrentRunspace);
-
-                var fullHelp = ps.AddScript($"(Get-Help -Name {commandName} -Full | Out-String)").Invoke<string>().FirstOrDefault();
-                if (!String.Equals(fullHelp, String.Empty, StringComparison.CurrentCultureIgnoreCase))
-                {
-                    WriteToAlternateScreenBuffer(fullHelp, String.IsNullOrEmpty(parameterName) ? null : parameterName);
-                }
-            }
-            else if (!String.IsNullOrEmpty(commandName) && !String.IsNullOrEmpty(parameterName))
-            {
-                System.Management.Automation.PowerShell ps = System.Management.Automation.PowerShell.Create(RunspaceMode.CurrentRunspace);
-                PSObject paramHelp = ps.AddScript($"(Get-Help -Name {commandName} -Parameter {parameterName})").Invoke<PSObject>().FirstOrDefault();
-
-                if (paramHelp != null)
-                {
-                    WriteParameterHelp(paramHelp);
-                }
-            }
-        }
-
-        private void WriteParameterHelp(dynamic helpContent)
-        {
-            char c = (char)0x1b;
-
-            string syntax = $"{c}[7m-{helpContent.name} <{helpContent.type.name}>{c}[0m";
-            string desc = "DESC: " + helpContent.Description[0].Text;
-            string details = $"Required: {helpContent.required}, Position: {helpContent.position}, Default Value: {helpContent.defaultValue}, Pipeline Input: {helpContent.pipelineInput}, WildCard: {helpContent.globbing}";
-
-            Collection<string> helpBlock = new Collection<string> {
-                String.Empty,
-                syntax,
-                String.Empty,
-                desc,
-                details
-            };
-
-            var bufferWidth = _console.BufferWidth;
-            var colWidth = Math.Min(helpBlock.Max(s => LengthInBufferCells(s)) + 2, bufferWidth);
-            int columns = 1;
-
-            var dynHelp = new DynamicHelp
-            {
-                Singleton = this,
-                ColumnWidth = colWidth,
-                Columns = columns,
-                Rows = (helpBlock.Count + columns - 1) / columns,
-                HelpItems = helpBlock
-            };
-
-            dynHelp.SaveCursor();
-            dynHelp.DrawHelpBlock(dynHelp);
-
-            Console.ReadKey(true);
-
-            dynHelp.Clear();
-            dynHelp.RestoreCursor();
-        }
-
-        private void WriteToAlternateScreenBuffer(string helpContent, string parameterName)
-        {
-            char c = (char) 0x1b;
-            string startAltBuffer = $"{c}[?1049h";
-            string endAltBuffer = $"{c}[?1049l";
-            string reverseColorStart = $"{c}[7m";
-            string reverseColorEnd = $"{c}[0m";
-
-            var contentAsArray = helpContent.Split(Environment.NewLine);
-
-            int startLine = 0;
-            int bufferHeight = Math.Min(Console.WindowHeight,Console.BufferHeight);
-            bool moved = true;
-
-            if (!String.IsNullOrEmpty(parameterName))
-            {
-                string upper = parameterName[0].ToString().ToUpperInvariant();
-                string lower = parameterName[0].ToString().ToLowerInvariant();
-                string remainingString = parameterName.Substring(1);
-                string pattern = $"-[{upper}|{lower}]{remainingString} [<|\\[]";
-
-                for(int i = 0; i < contentAsArray.Length; i++)
-                {
-                    if (System.Text.RegularExpressions.Regex.IsMatch(contentAsArray[i], pattern))
-                    {
-                        startLine = i;
-                    }
-                }
-            }
-
-            string pagerMessage = String.Concat(
-                    reverseColorStart,
-                    "Up:",
-                    reverseColorEnd,
-                    "↑ ",
-                    reverseColorStart,
-                    "Down:",
-                    reverseColorEnd,
-                    "↓ ",
-                    reverseColorStart,
-                    "Quit:",
-                    reverseColorEnd,
-                    "Q :");
-
-            Console.Write(startAltBuffer);
-
-            do
-            {
-                if (moved) {
-                    Console.Clear();
-                    int offset = GetMultilineOffset(contentAsArray, startLine);
-                    string content = String.Join(Environment.NewLine, contentAsArray.Skip(startLine).Take(bufferHeight - offset));
-                    Console.WriteLine(content);
-                    Console.Write(pagerMessage);
-                }
-
-                var pressed = Console.ReadKey(intercept: true);
-
-                if (pressed.Key == ConsoleKey.UpArrow) {
-                    if (startLine > 0) {
-                        startLine--;
-                        moved = true;
-                    }
-                }
-                else if (pressed.Key == ConsoleKey.DownArrow) {
-                    if ((startLine + bufferHeight) < contentAsArray.Count()) {
-                        startLine++;
-                        moved = true;
-                    }
-                }
-                else if (pressed.Key == ConsoleKey.Q) {
-                    break;
-                }
-                else {
-                    moved = false;
-                }
-
-            } while (true);
-
-            Console.Write(endAltBuffer);
-        }
-
-        private int GetMultilineOffset(string[] contentAsArray, int startLine)
-        {
-            int bufferHeight = Math.Min(Console.WindowHeight,Console.BufferHeight);
-            int contentTotalLines = contentAsArray.Count();
-            int endLine = startLine + bufferHeight;
-            int bufferWidth = Console.BufferWidth;
-
-            int offset = 0;
-
-            for(int i = startLine; i < endLine; i++)
-            {
-                int lineLength = contentAsArray[i].Length;
-                if (lineLength > bufferWidth)
-                {
-                    offset += lineLength / bufferWidth;
-                }
-            }
-
-            return offset;
         }
 
         private void CompleteImpl(bool menuSelect)
@@ -625,253 +421,6 @@ namespace Microsoft.PowerShell
                 s = s.Substring(0, newlineIndex) + "...";
             }
             return s;
-        }
-
-        private class DynamicHelp
-        {
-            internal PSConsoleReadLine Singleton;
-            internal int Top;
-
-            internal int PreviousTop;
-            internal int ColumnWidth;
-            internal int BufferLines;
-            internal int Rows;
-            internal int Columns;
-            internal int ToolTipLines;
-            internal Collection<string> HelpItems;
-            //internal CompletionResult CurrentMenuItem => MenuItems[CurrentSelection];
-            internal int CurrentSelection;
-
-            public void DrawHelpBlock(DynamicHelp previousMenu, bool menuSelect = true)
-            {
-                IConsole console = Singleton._console;
-
-                // Move cursor to the start of the first line after our input.
-                var bufferEndPoint = Singleton.ConvertOffsetToPoint(Singleton._buffer.Length);
-                console.SetCursorPosition(bufferEndPoint.X, bufferEndPoint.Y);
-                // Top must be initialized before calling AdjustForPossibleScroll, otherwise
-                // on the last line of the buffer, the scroll operation causes Top to point
-                // past the buffer, which in turn causes the menu to be printed twice.
-                this.Top = bufferEndPoint.Y + 1;
-                AdjustForPossibleScroll(1);
-                MoveCursorDown(1);
-
-                var bufferWidth = console.BufferWidth;
-                var columnWidth = this.ColumnWidth;
-
-                var items = this.HelpItems;
-                for (var row = 0; row < this.Rows; row++)
-                {
-                    var cells = 0;
-                    for (var col = 0; col < this.Columns; col++)
-                    {
-                        var index = row + (this.Rows * col);
-                        if (index >= items.Count)
-                        {
-                            break;
-                        }
-                        console.Write(GetHelpItem(items[index], columnWidth));
-                        cells += columnWidth;
-                    }
-
-                    // Make sure we always write out exactly 1 buffer width to erase anything
-                    // from a previous menu.
-                    if (cells < bufferWidth)
-                    {
-                        // 'BlankRestOfLine' erases rest of the current line, but the cursor is not moved.
-                        console.BlankRestOfLine();
-                    }
-
-                    // Explicit newline so consoles see each row as distinct lines, but skip the
-                    // last line so we don't scroll.
-                    if (row != (this.Rows - 1) || !menuSelect) {
-                        AdjustForPossibleScroll(1);
-                        MoveCursorDown(1);
-                    }
-                }
-
-                if (previousMenu != null)
-                {
-                    if (Rows < previousMenu.Rows + previousMenu.ToolTipLines)
-                    {
-                        // Rest of the current line was erased, but the cursor was not moved to the next line.
-                        if (console.CursorLeft != 0)
-                        {
-                            // There are lines from the previous rendering that need to be cleared,
-                            // so we are sure there is no need to scroll.
-                            MoveCursorDown(1);
-                        }
-
-                        Singleton.WriteBlankLines(previousMenu.Rows + previousMenu.ToolTipLines - Rows);
-                    }
-                }
-
-                // if the menu has moved, we need to clear the lines under it
-                if (bufferEndPoint.Y < PreviousTop)
-                {
-                    console.BlankRestOfLine();
-                    Singleton.WriteBlankLines(PreviousTop - bufferEndPoint.Y);
-                }
-
-                PreviousTop = bufferEndPoint.Y;
-
-                if (menuSelect)
-                {
-                    RestoreCursor();
-                    console.CursorVisible = true;
-                }
-            }
-
-            public void Clear()
-            {
-                WriteBlankLines(Top, Rows + ToolTipLines);
-            }
-
-            /* public void UpdateMenuSelection(int selectedItem, bool select, bool showTooltips, string toolTipColor)
-            {
-                var console = Singleton._console;
-                var menuItem = MenuItems[selectedItem];
-                var listItem = menuItem.ListItemText;
-
-                string toolTip = null;
-                if (showTooltips)
-                {
-                    toolTip = menuItem.ToolTip.Trim();
-
-                    // Don't bother showing the tooltip if it doesn't add information.
-                    showTooltips = !string.IsNullOrWhiteSpace(toolTip)
-                        && !string.Equals(toolTip, listItem, StringComparison.OrdinalIgnoreCase)
-                        && !string.Equals(toolTip, menuItem.CompletionText, StringComparison.OrdinalIgnoreCase);
-                }
-
-                // We'll use one blank line to set the tooltip apart from the menu,
-                // and there will be at least 1 line in the tooltip, possibly more.
-                var toolTipLines = 2;
-                if (showTooltips)
-                {
-                    // Determine if showing the tooltip would scroll the top of our buffer off the screen.
-
-                    int lineLength = 0;
-                    for (var i = 0; i < toolTip.Length; i++)
-                    {
-                        char c = toolTip[i];
-                        if (c == '\r' && i < toolTip.Length && toolTip[i+1] == '\n')
-                        {
-                            // Skip the newline, but handle LF, CRLF, and CR.
-                            i += 1;
-                        }
-
-                        if (c == '\r' || c == '\n')
-                        {
-                            toolTipLines += 1;
-                            lineLength = 0;
-                        }
-                        else
-                        {
-                            lineLength += 1;
-                            if (lineLength == console.BufferWidth)
-                            {
-                                toolTipLines += 1;
-                                lineLength = 0;
-                            }
-                        }
-                    }
-
-                    // The +1 is for the blank line between the menu and tooltips.
-                    if (BufferLines + Rows + toolTipLines + 1 > console.WindowHeight)
-                    {
-                        showTooltips = false;
-                    }
-                }
-
-                SaveCursor();
-
-                var row = Top + selectedItem % Rows;
-                var col = ColumnWidth * (selectedItem / Rows);
-
-                console.SetCursorPosition(col, row);
-
-                if (select) console.Write(Singleton.Options._selectionColor);
-                console.Write(GetMenuItem(listItem, ColumnWidth));
-                if (select) console.Write("\x1b[0m");
-
-                ToolTipLines = 0;
-                if (showTooltips)
-                {
-                    Debug.Assert(select, "On unselect, caller must clear the tooltip");
-                    console.SetCursorPosition(0, Top + Rows - 1);
-                    // Move down 2 so we have 1 blank line between the menu and buffer.
-                    AdjustForPossibleScroll(toolTipLines);
-                    MoveCursorDown(2);
-                    console.Write(toolTipColor);
-                    console.Write(toolTip);
-                    ToolTipLines = toolTipLines;
-
-                    console.Write("\x1b[0m");
-                }
-
-                RestoreCursor();
-            }*/
-
-            /* public void MoveRight()    => CurrentSelection = Math.Min(CurrentSelection + Rows, HelpItems.Count - 1);
-            public void MoveLeft()     => CurrentSelection = Math.Max(CurrentSelection - Rows, 0);
-            public void MoveUp()       => CurrentSelection = Math.Max(CurrentSelection - 1, 0);
-            public void MoveDown()     => CurrentSelection = Math.Min(CurrentSelection + 1, HelpItems.Count - 1);
-            public void MovePageDown() => CurrentSelection = Math.Min(CurrentSelection + Rows - (CurrentSelection % Rows) - 1,
-                                                                      HelpItems.Count - 1);
-            public void MovePageUp()   => CurrentSelection = Math.Max(CurrentSelection - (CurrentSelection % Rows), 0);
-
-            public void MoveN(int n)
-            {
-                CurrentSelection = (CurrentSelection + n) % MenuItems.Count;
-                if (CurrentSelection < 0)
-                {
-                    CurrentSelection += MenuItems.Count;
-                }
-            }
-            */
-
-            private void MoveCursorDown(int cnt)
-            {
-                IConsole console = Singleton._console;
-                while (cnt-- > 0)
-                {
-                    console.Write("\n");
-                }
-            }
-
-
-            private void AdjustForPossibleScroll(int cnt)
-            {
-                IConsole console = Singleton._console;
-                var scrollCnt = console.CursorTop + cnt + 1 - console.BufferHeight;
-                if (scrollCnt > 0)
-                {
-                    Top -= scrollCnt;
-                    _singleton._initialY -= scrollCnt;
-                    _savedCursorTop -= scrollCnt;
-                }
-            }
-
-            public void WriteBlankLines(int top, int count)
-            {
-                SaveCursor();
-                Singleton._console.SetCursorPosition(0, top);
-                Singleton.WriteBlankLines(count);
-                RestoreCursor();
-            }
-
-            private int _savedCursorLeft;
-            private int _savedCursorTop;
-
-            public void SaveCursor()
-            {
-                IConsole console = Singleton._console;
-                _savedCursorLeft = console.CursorLeft;
-                _savedCursorTop = console.CursorTop;
-            }
-
-            public void RestoreCursor() => Singleton._console.SetCursorPosition(_savedCursorLeft, _savedCursorTop);
         }
 
         private class Menu
